backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
faiss-cpu>=1.7.4
firebase-admin>=6.2.0
google-cloud-storage>=2.12.0
python-dotenv>=1.0.0
langchain>=0.0.335
numpy>=1.24.4
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=0.8.11
chardet>=5.1.0
google-generativeai>=0.3.1
